+++ cpp/src/decisiontree/kernels/quantile.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2019, NVIDIA CORPORATION.
+ * Copyright (c) 2019, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,15 +15,9 @@
  */
 
 #pragma once
-
-#include <distance/distance.h>
-#include <common/cumlHandle.hpp>
-
-namespace ML {
-namespace Metrics {
-
-template <typename math_t, MLCommon::Distance::DistanceType distance_type>
-double trustworthiness_score(const cumlHandle& h, math_t* X, math_t* X_embedded,
-                             int n, int m, int d, int n_neighbors);
-}
-}  // namespace ML
+#include "../memory.h"
+template <typename T, typename L>
+void preprocess_quantile(const T *data, const unsigned int *rowids,
+                         const int n_sampled_rows, const int ncols,
+                         const int rowoffset, const int nbins,
+                         std::shared_ptr<TemporaryMemory<T, L>> tempmem);