+++ cpp/test/sg/rf_test.cu
@@ -39,6 +39,7 @@ struct RfInputs {
   int n_bins;
   int split_algo;
   int min_rows_per_node;
+  CRITERION split_criterion;
 };
 
 template <typename T>
@@ -47,14 +48,15 @@ template <typename T>
 }
 
 template <typename T>
-class RfTest : public ::testing::TestWithParam<RfInputs<T>> {
+class RfClassifierTest : public ::testing::TestWithParam<RfInputs<T>> {
  protected:
   void basicTest() {
     params = ::testing::TestWithParam<RfInputs<T>>::GetParam();
 
     DecisionTree::DecisionTreeParams tree_params(
       params.max_depth, params.max_leaves, params.max_features, params.n_bins,
-      params.split_algo, params.min_rows_per_node, params.bootstrap_features);
+      params.split_algo, params.min_rows_per_node, params.bootstrap_features,
+      params.split_criterion, false);
     RF_params rf_params(params.bootstrap, params.bootstrap_features,
                         params.n_trees, params.rows_sample, tree_params);
     //rf_params.print();
@@ -66,6 +68,8 @@ class RfTest : public ::testing::TestWithParam<RfInputs<T>> {
     int data_len = params.n_rows * params.n_cols;
     allocate(data, data_len);
     allocate(labels, params.n_rows);
+    allocate(predicted_labels, params.n_inference_rows);
+
     cudaStream_t stream;
     CUDA_CHECK(cudaStreamCreate(&stream));
 
@@ -89,18 +93,21 @@ class RfTest : public ::testing::TestWithParam<RfInputs<T>> {
         labels_map.size());
 
     CUDA_CHECK(cudaStreamSynchronize(stream));
-    CUDA_CHECK(cudaStreamDestroy(stream));
 
     // Inference data: same as train, but row major
     int inference_data_len = params.n_inference_rows * params.n_cols;
     inference_data_h = {30.0, 10.0, 1.0, 20.0, 2.0, 10.0, 0.0, 40.0};
     inference_data_h.resize(inference_data_len);
+    allocate(inference_data_d, inference_data_len);
+    updateDevice(inference_data_d, inference_data_h.data(), data_len, stream);
 
     // Predict and compare against known labels
-    predicted_labels.resize(params.n_inference_rows);
-    RF_metrics tmp = score(handle, rf_classifier, inference_data_h.data(),
-                           labels_h.data(), params.n_inference_rows,
-                           params.n_cols, predicted_labels.data(), false);
+    RF_metrics tmp =
+      score(handle, rf_classifier, inference_data_d, labels,
+            params.n_inference_rows, params.n_cols, predicted_labels, false);
+    CUDA_CHECK(cudaStreamSynchronize(stream));
+    CUDA_CHECK(cudaStreamDestroy(stream));
+
     accuracy = tmp.accuracy;
   }
 
@@ -112,16 +119,17 @@ class RfTest : public ::testing::TestWithParam<RfInputs<T>> {
     inference_data_h.clear();
     labels_h.clear();
     labels_map.clear();
-    predicted_labels.clear();
 
     CUDA_CHECK(cudaFree(labels));
+    CUDA_CHECK(cudaFree(predicted_labels));
     CUDA_CHECK(cudaFree(data));
+    CUDA_CHECK(cudaFree(inference_data_d));
     delete rf_classifier;
   }
 
  protected:
   RfInputs<T> params;
-  T* data;
+  T *data, *inference_data_d;
   int* labels;
   std::vector<T> inference_data_h;
   std::vector<int> labels_h;
@@ -131,33 +139,153 @@ class RfTest : public ::testing::TestWithParam<RfInputs<T>> {
   rfClassifier<T>* rf_classifier;
   float accuracy = -1.0f;  // overriden in each test SetUp and TearDown
 
-  std::vector<int> predicted_labels;
+  int* predicted_labels;
 };
 
-const std::vector<RfInputs<float>> inputsf2 = {
-  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST,
-   2},  // single tree forest, bootstrap false, unlimited depth, 4 bins
-  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST,
-   2},  // single tree forest, bootstrap false, depth of 8, 4 bins
-  {4, 2, 10, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST,
-   2},  //forest with 10 trees, all trees should produce identical predictions (no bootstrapping or column subsampling)
-  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::HIST,
-   2},  //forest with 10 trees, with bootstrap and column subsampling enabled, 3 bins
-  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::GLOBAL_QUANTILE,
-   2}  //forest with 10 trees, with bootstrap and column subsampling enabled, 3 bins, different split algorithm
+//-------------------------------------------------------------------------------------------------------------------------------------
+
+template <typename T>
+class RfRegressorTest : public ::testing::TestWithParam<RfInputs<T>> {
+ protected:
+  void basicTest() {
+    params = ::testing::TestWithParam<RfInputs<T>>::GetParam();
+
+    DecisionTree::DecisionTreeParams tree_params(
+      params.max_depth, params.max_leaves, params.max_features, params.n_bins,
+      params.split_algo, params.min_rows_per_node, params.bootstrap_features,
+      params.split_criterion, false);
+    RF_params rf_params(params.bootstrap, params.bootstrap_features,
+                        params.n_trees, params.rows_sample, tree_params);
+    //rf_params.print();
+
+    //--------------------------------------------------------
+    // Random Forest
+    //--------------------------------------------------------
+
+    int data_len = params.n_rows * params.n_cols;
+    allocate(data, data_len);
+    allocate(labels, params.n_rows);
+    allocate(predicted_labels, params.n_inference_rows);
+    cudaStream_t stream;
+    CUDA_CHECK(cudaStreamCreate(&stream));
+
+    // Populate data (assume Col major)
+    std::vector<T> data_h = {0.0, 0.0, 0.0, 0.0, 10.0, 20.0, 30.0, 40.0};
+    data_h.resize(data_len);
+    updateDevice(data, data_h.data(), data_len, stream);
+
+    // Populate labels
+    labels_h = {1.0, 2.0, 3.0, 4.0};
+    labels_h.resize(params.n_rows);
+    updateDevice(labels, labels_h.data(), params.n_rows, stream);
+
+    rf_regressor = new typename rfRegressor<T>::rfRegressor(rf_params);
+
+    cumlHandle handle;
+    handle.setStream(stream);
+
+    fit(handle, rf_regressor, data, params.n_rows, params.n_cols, labels);
+
+    CUDA_CHECK(cudaStreamSynchronize(stream));
+
+    // Inference data: same as train, but row major
+    int inference_data_len = params.n_inference_rows * params.n_cols;
+    inference_data_h = {0.0, 10.0, 0.0, 20.0, 0.0, 30.0, 0.0, 40.0};
+    inference_data_h.resize(inference_data_len);
+    allocate(inference_data_d, inference_data_len);
+    updateDevice(inference_data_d, inference_data_h.data(), data_len, stream);
+
+    // Predict and compare against known labels
+    RF_metrics tmp =
+      score(handle, rf_regressor, inference_data_d, labels,
+            params.n_inference_rows, params.n_cols, predicted_labels, false);
+    CUDA_CHECK(cudaStreamSynchronize(stream));
+    CUDA_CHECK(cudaStreamDestroy(stream));
+
+    mse = tmp.mean_squared_error;
+  }
+
+  void SetUp() override { basicTest(); }
+
+  void TearDown() override {
+    mse = -1.0f;  // reset mse
+    inference_data_h.clear();
+    labels_h.clear();
+
+    CUDA_CHECK(cudaFree(labels));
+    CUDA_CHECK(cudaFree(predicted_labels));
+    CUDA_CHECK(cudaFree(data));
+    CUDA_CHECK(cudaFree(inference_data_d));
+    delete rf_regressor;
+  }
+
+ protected:
+  RfInputs<T> params;
+  T *data, *inference_data_d;
+  T* labels;
+  std::vector<T> inference_data_h;
+  std::vector<T> labels_h;
+
+  rfRegressor<T>* rf_regressor;
+  float mse = -1.0f;  // overriden in each test SetUp and TearDown
+
+  T* predicted_labels;
 };
+//-------------------------------------------------------------------------------------------------------------------------------------
 
-const std::vector<RfInputs<double>> inputsd2 = {  // Same as inputsf2
-  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2},
-  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2},
-  {4, 2, 10, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2},
-  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::HIST, 2},
+const std::vector<RfInputs<float>> inputsf2_clf = {
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::
+     GINI},  // single tree forest, bootstrap false, unlimited depth, 4 bins
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::GINI},  // single tree forest, bootstrap false, depth of 8, 4 bins
+  {4, 2, 10, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::
+     GINI},  //forest with 10 trees, all trees should produce identical predictions (no bootstrapping or column subsampling)
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::HIST, 2,
+   CRITERION::
+     GINI},  //forest with 10 trees, with bootstrap and column subsampling enabled, 3 bins
   {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::GLOBAL_QUANTILE,
-   2}};
+   2,
+   CRITERION::
+     CRITERION_END},  //forest with 10 trees, with bootstrap and column subsampling enabled, 3 bins, different split algorithm
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 10, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::GLOBAL_QUANTILE,
+   2, CRITERION::ENTROPY}};
 
-typedef RfTest<float> RfTestF;
-TEST_P(RfTestF, Fit) {
-  //rf_classifier->print_rf_detailed(); // Prints all trees in the forest. Leaf nodes use the remapped values from labels_map.
+const std::vector<RfInputs<double>> inputsd2_clf = {  // Same as inputsf2_clf
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::GINI},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::GINI},
+  {4, 2, 10, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::GINI},
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::HIST, 2,
+   CRITERION::GINI},
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::GLOBAL_QUANTILE,
+   2, CRITERION::CRITERION_END},
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 10, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::HIST, 2,
+   CRITERION::ENTROPY},
+  {4, 2, 10, 0.8f, 0.8f, 4, 8, -1, true, false, 3, SPLIT_ALGO::GLOBAL_QUANTILE,
+   2, CRITERION::ENTROPY}};
+
+typedef RfClassifierTest<float> RfClassifierTestF;
+TEST_P(RfClassifierTestF, Fit) {
+  //rf_classifier
+  //  ->print_rf_detailed();  // Prints all trees in the forest. Leaf nodes use the remapped values from labels_map.
   if (!params.bootstrap && (params.max_features == 1.0f)) {
     ASSERT_TRUE(accuracy == 1.0f);
   } else {
@@ -165,8 +293,8 @@ TEST_P(RfTestF, Fit) {
   }
 }
 
-typedef RfTest<double> RfTestD;
-TEST_P(RfTestD, Fit) {
+typedef RfClassifierTest<double> RfClassifierTestD;
+TEST_P(RfClassifierTestD, Fit) {
   if (!params.bootstrap && (params.max_features == 1.0f)) {
     ASSERT_TRUE(accuracy == 1.0f);
   } else {
@@ -174,8 +302,63 @@ TEST_P(RfTestD, Fit) {
   }
 }
 
-INSTANTIATE_TEST_CASE_P(RfTests, RfTestF, ::testing::ValuesIn(inputsf2));
+INSTANTIATE_TEST_CASE_P(RfClassifierTests, RfClassifierTestF,
+                        ::testing::ValuesIn(inputsf2_clf));
 
-INSTANTIATE_TEST_CASE_P(RfTests, RfTestD, ::testing::ValuesIn(inputsd2));
+INSTANTIATE_TEST_CASE_P(RfClassifierTests, RfClassifierTestD,
+                        ::testing::ValuesIn(inputsd2_clf));
+
+typedef RfRegressorTest<float> RfRegressorTestF;
+TEST_P(RfRegressorTestF, Fit) {
+  //rf_regressor->print_rf_detailed(); // Prints all trees in the forest.
+  if (!params.bootstrap && (params.max_features == 1.0f)) {
+    ASSERT_TRUE(mse == 0.0f);
+  } else {
+    ASSERT_TRUE(mse <= 0.2f);
+  }
+}
+
+typedef RfRegressorTest<double> RfRegressorTestD;
+TEST_P(RfRegressorTestD, Fit) {
+  if (!params.bootstrap && (params.max_features == 1.0f)) {
+    ASSERT_TRUE(mse == 0.0f);
+  } else {
+    ASSERT_TRUE(mse <= 0.2f);
+  }
+}
+
+const std::vector<RfInputs<float>> inputsf2_reg = {
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::MSE},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::MSE},
+  {4, 2, 5, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::
+     CRITERION_END},  // CRITERION_END uses the default criterion (GINI for classification, MSE for regression)
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::MAE},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::GLOBAL_QUANTILE,
+   2, CRITERION::MAE},
+  {4, 2, 5, 1.0f, 1.0f, 4, 8, -1, true, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::CRITERION_END}};
+
+const std::vector<RfInputs<double>> inputsd2_reg = {  // Same as inputsf2_reg
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::MSE},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::MSE},
+  {4, 2, 5, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::CRITERION_END},
+  {4, 2, 1, 1.0f, 1.0f, 4, -1, -1, false, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::MAE},
+  {4, 2, 1, 1.0f, 1.0f, 4, 8, -1, false, false, 4, SPLIT_ALGO::GLOBAL_QUANTILE,
+   2, CRITERION::MAE},
+  {4, 2, 5, 1.0f, 1.0f, 4, 8, -1, true, false, 4, SPLIT_ALGO::HIST, 2,
+   CRITERION::CRITERION_END}};
+
+INSTANTIATE_TEST_CASE_P(RfRegressorTests, RfRegressorTestF,
+                        ::testing::ValuesIn(inputsf2_reg));
+INSTANTIATE_TEST_CASE_P(RfRegressorTests, RfRegressorTestD,
+                        ::testing::ValuesIn(inputsd2_reg));
 
 }  // end namespace ML